requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
azure-cognitiveservices-speech==1.34.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.109.0
frozenlist==1.4.1
h11==0.14.0
httptools==0.6.1
idna==3.6
multidict==6.0.4
pydantic==2.5.3
pydantic-core==2.14.6
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
sse-starlette==1.8.2
starlette==0.35.1
typing-extensions==4.9.0
urllib3==2.1.0
uvicorn==0.26.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
